simplify slide seeking logic
